backend/pyproject.toml
[tool.poetry]
name = "portfolio-backend"
version = "0.1.0"
description = ""
authors = ["okuzmenko31 <olegkuzmenko2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
starlette = "0.27.0"
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
greenlet = "^2.0.2"
jinja2 = "^3.1.2"
fastapi-mail = "^1.4.1"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
cryptography = "^41.0.2"
pydantic = "2.1.1"
python-multipart = "^0.0.6"
uvicorn = "^0.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
